Move CellSuccessData tests to their own file
